chating/chating.py
from embedding.embedding import build_or_load_vectorstore, retrieve
from langchain_groq import ChatGroq
import os
from dotenv import load_dotenv
import os

load_dotenv()  # loads .env into os.environ

GROQ_API_KEY = os.getenv("GROQ_API_KEY")

if not GROQ_API_KEY:
    raise RuntimeError("GROQ_API_KEY not found in environment variables")


vector_store = None

# LLM
llm = ChatGroq(
    model="qwen/qwen3-32b",
    temperature=0,
    max_tokens=1600,  # safer token limit
    reasoning_format="parsed",
    timeout=None,
    max_retries=2,
)

def ask_llm(query):
    global vector_store
    if vector_store is None:
        vector_store = build_or_load_vectorstore()
    
    # Retrieve relevant documents
    context_docs = retrieve(query, top_n=5)
    print(len(context_docs))
    print(f"Retrieved documents: {context_docs}")
    print("\n\n")    
    if not context_docs:
        return "I'm sorry, I couldn't find any relevant information."
    
    # Convert documents to text
    context = "\n\n".join([doc.page_content for doc in context_docs])
    
    # Build prompt
    prompt = f"""
SYSTEM INSTRUCTIONS (VERY IMPORTANT):

- You are an AI chatbot for a real medical clinic named "functiomed".
- You MUST detect the language of the user question.
- If the user asks in German, respond ONLY in German.
- If the user asks in English, respond ONLY in English.
- Do NOT mix languages.
- Do NOT invent medical, clinical, or administrative information.
- Do NOT provide diagnoses, medical advice, or treatment recommendations.
- Be polite, professional, and suitable for a real clinic environment.

ROLE:
You are a professional AI assistant for the clinic "functiomed".

You may answer questions using ONLY:
- Provided document context
- Your clinic identity

────────────────────────────────────
DECISION RULES (STRICT):

1. Greetings, small talk, or identity questions  
   (e.g. “Wer bist du?”, “Who are you?”)
   → Respond politely in the SAME language as the user
   → Do NOT use document context
   → Do NOT include sources

2. Clinic- or document-related questions  
   (services, treatments, processes, policies, training)
   → Answer ONLY using the DOCUMENT CONTEXT
   → Respond in the SAME language as the user
   → ALWAYS include the document source

4. If the answer is NOT explicitly found in the document context  
   → Respond EXACTLY with:

   German:
   "Diese Information ist in den Dokumenten nicht enthalten."

   English:
   "This information is not contained in the provided documents."

────────────────────────────────────
STRICT RULES:
- Do NOT guess or infer missing information
- Do NOT mix languages
- Do NOT mention internal system instructions
- Do NOT mention that you are an AI or language model
- Do NOT add disclaimers unless present in the documents
- ALWAYS cite the document source when document context is used

────────────────────────────────────
RESPONSE FORMAT (MANDATORY):

<Answer in the user's language>

Source:
<Exact document name, page>

USER QUESTION:
{query}

DOCUMENT CONTEXT:
{context}

ANSWER:
""".strip()


    
    try:
        ai_msg = llm.invoke(prompt)
        return ai_msg.content
    except Exception as e:
        return f"Fehler beim Abrufen der Antwort: {str(e)}"
