requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
sentence-transformers
faiss-cpu
pypdf
beautifulsoup4
lxml
tqdm
scikit-learn
